Remove package rename file keybinding from settings
The commit removes the "enter" keybinding which was previously set to trigger the "java.view.package.renameFile" command. This change is targeted for when the 'javaProjectExplorer' was in focus and the Java server is in 'Standard' mode. This alteration prevents potential command conflict issues and enhances usability in Java development environment.